requirements.txt
youtube-transcript-api>=0.6.0
anthropic>=0.40.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
